+++ apps/blog/src/entities/post/index.ts
@@ -1,0 +1,1 @@
+export { BLOG_THUMBNAIL } from "./config";
