+++ apps/blog/src/entities/project/index.ts
@@ -1,0 +1,1 @@
+export { PROJECTS_DATA } from "./config";
